FIX Rename ardupilot_methodic_configurator.py to __main__.py
