+++ setup.py
@@ -26,7 +26,7 @@ PRJ_URL = "https://github.com/ArduPilot/MethodicConfigurator"
 for file in extra_scripts:
     os.chmod(file, 0o755)  # noqa: S103
 
-os.chmod("ardupilot_methodic_configurator/ardupilot_methodic_configurator.py", 0o755)  # noqa: S103
+os.chmod("ardupilot_methodic_configurator/__main__.py", 0o755)  # noqa: S103
 
 # Read the long description from the README file
 with open("README.md", encoding="utf-8") as f:
